Dhall.Typed

Conversion

To

toTypedKind :: TermCtx us vs -> Expr () X -> Maybe DKind

Convert an untyped Dhall expression representing a kind into a DKind.

Will fail if:

The Dhall expression does not represent a kind

Any kind variables are involved. Kind variables are not yet supported! But there is no fundamental reason why they wouldn't be; they just have not been implemented yet.

Will behave unpredictably if the Dhall expression does not typecheck within Dhall itself.

toTypedType

Arguments

:: SDKind k

-> TermCtx us vs
kinds and types of free variables

-> Expr () X

-> Maybe (DType us k)

Convert an untyped Dhall expression into a typed one representing a Dhall type of a desired kind.

Will fail if:

The Dhall expression does not represent a type

The kind does not match

Any kind variables are involved. Kind variables are not yet supported! But there is no fundamental reason why they wouldn't be; they just have not been implemented yet.

Will behave unpredictably if the Dhall expression does not typecheck within Dhall itself.

toTypedTerm

Arguments

:: SDType us Type a

-> TermCtx us vs
kinds and types of free variables

-> Expr () X

-> Maybe (DTerm us vs a)

Convert an untyped Dhall expression into a typed one representing a Dhall term of a desired type.

Will fail if:

The Dhall expression does not represent a term

The type does not match

Any kind variables are involved. Kind variables are not yet supported! But there is no fundamental reason why they wouldn't be; they just have not been implemented yet.

Will behave unpredictably if the Dhall expression does not typecheck within Dhall itself.

toSomeTerm :: TermCtx us vs -> Expr () X -> Maybe (SomeTerm us vs)

Convert an untyped Dhall expression representing a term into a typed one, also determining the type in the process.

Will fail if:

The Dhall expression does not represent a term

Any kind variables are involved. Kind variables are not yet supported! But there is no fundamental reason why they wouldn't be; they just have not been implemented yet.

Will behave unpredictably if the Dhall expression does not typecheck.

From

fromTypedKind :: DKind -> Expr s X

fromTypedType :: DType us k -> Expr s X

fromTypedTerm :: DTerm us vs a -> Expr s X

Typechecking

typeOfExpr :: TermCtx us vs -> Expr () X -> Maybe (DType us Type)

Find the DType corresponding to the type of the Dhall expression representing a term.

Will fail if:

The Dhall expression does not represent a term

Any kind variables are involved. Kind variables are not yet supported! But there is no fundamental reason why they wouldn't be; they just have not been implemented yet.

Will behave unpredictably if the Dhall expression does not typecheck within Dhall itself.

Context

data TermCtx us :: [DType us Type] -> Type where

Constructors

TCtxNil :: TermCtx '[] '[]

ConsSort
Not supported, but might be one day.

Fields

:: Text

-> Expr () X

we have to keep original sort, since we can't synthesize it

-> TermCtx us vs

-> TermCtx us vs

ConsKind :: Text -> SDKind u -> TermCtx us vs -> TermCtx (u ': us) (MapShift u us vs)

ConsType :: Text -> SDType us Type v -> TermCtx us vs -> TermCtx us (v ': vs)

ctxKinds :: TermCtx us vs -> Prod SDKind us

ctxTypes :: TermCtx us vs -> Prod (SDType us Type) vs

toContext :: TermCtx us vs -> Context (Expr () X)

data TermCtxItem us :: [DType us Type] -> Type where

Constructors

TCISort :: Expr () X -> TermCtxItem us vs

TCIKind :: Index us u -> SDKind u -> TermCtxItem us vs

TCIType :: Index vs v -> SDType us Type v -> TermCtxItem us vs

lookupCtx :: Text -> Integer -> TermCtx us vs -> Maybe (TermCtxItem us vs)

Convenience

kindcheckType :: forall a b us. SDKindI b => SDKind a -> DType us b -> Maybe (DType us a)

typecheckTerm :: forall us vs a b. SDTypeI us Type b => Prod SDKind us -> SDType us Type a -> DTerm us vs b -> Maybe (DTerm us vs a)

Samples

ident :: DTerm us vs (Pi SType (TVar IZ :-> TVar IZ))

The identity function, encoded as a DTerm. Provided as an example.

konst :: DTerm us vs (Pi SType (Pi SType (TVar (IS IZ) :-> (TVar IZ :-> TVar (IS IZ)))))

The constant function, encoded as a DTerm. Provided as an example. All of the multi-Pi functions here require the typechecker plugin.

konst' :: DTerm us vs (Pi SType (TVar IZ :-> Pi SType (TVar IZ :-> TVar (IS IZ))))

The constant function with flipped parameter order, encoded as a DTerm. Provided as an example.

konst3 :: DTerm us vs (Pi SType (Pi SType (Pi SType (TVar (IS (IS IZ)) :-> (TVar (IS IZ) :-> (TVar IZ :-> TVar (IS (IS IZ))))))))

The constant function with three inputs, encoded as a DTerm. Provided as an example.

konst4 :: DTerm us vs (Pi SType (Pi SType (Pi SType (Pi SType (TVar (IS (IS (IS IZ))) :-> (TVar (IS (IS IZ)) :-> (TVar (IS IZ) :-> (TVar IZ :-> TVar (IS (IS (IS IZ)))))))))))

The constant function with four inputs, encoded as a DTerm. Provided as an example.

natBuild :: DTerm us vs (Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)) :-> Natural)

Natural/build, encoded as a DTerm. Provided as an example.

listBuild :: DTerm us vs (Pi SType (Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)) :-> (List :$ TVar IZ)))

List/build, encoded as a DTerm. Provided as an example.