Dhall.Typed.Core

Expression

Sorts

data DSort :: Type where

Represents the possible sorts encountered in Dhall.

Note that this implementation allows records of kinds with sorts other than Kind, so { foo : Kind -> Kind } would typecheck, even though normal Dhall forbids this.

Constructors

Kind :: DSort

(:*>) :: DSort -> DSort -> DSort

SRecord :: AggType () () ls as -> Prod (Const DSort) as -> DSort

SUnion :: AggType () () ls as -> Prod (Const DSort) as -> DSort

Kinds

data KPrim :: [DSort] -> DSort -> Type where

Constructors

KRecord :: AggType DSort Kind ls as -> KPrim as Kind

KUnion :: AggType DSort Kind ls as -> KPrim as Kind

KRecordLit :: RecordVal () (Const DSort) () ls ks at bs as -> KPrim (Map GetConstSym0 as) (SRecord at bs)

KUnionLit :: UnionVal () (Const DSort) () ls ks at bs (Const a) -> KPrim '[a] (SRecord at bs)

data DKind :: [DSort] -> DSort -> Type where

Represents the possible types encountered in Dhall. A value of type

DKind '[r, s] t

Describes a kind of sort t with possible kind variables of sorts r and s.

Something of type DKind '[] t is a kind of sort t with no free variables.

Constructors

KVar :: Index ts a -> DKind ts a

KLam :: SDSort t -> DKind (t ': ts) a -> DKind ts (t :*> a)

KApp :: DKind ts (a :*> b) -> DKind ts a -> DKind ts b

(:~>) :: DKind ts Kind -> DKind ts Kind -> DKind ts Kind infixr 1

KPi :: SDSort t -> DKind (t ': ts) a -> DKind ts a

Type :: DKind ts Kind

KP :: KPrim as a -> Prod (DKind ts) as -> DKind ts a

data SomeKind :: [DSort] -> Type where

Constructors

SomeKind :: SDSort a -> DKind ts a -> SomeKind ts

type (:~>) a b = a :~> b infixr 1

type family KShift ts ps a b (ins :: Insert ts ps a) (x :: DKind ts b) :: DKind ps b where ...

Shift all kind variables in a kind expression of sort b to account for a new bound variable of sort a, to be inserted at the position indicated by the Insert.

Types

data TPrim ts :: [DKind ts Kind] -> DKind ts Kind -> Type where

Primitives of Dhall types, built into the language.

Constructors

Bool :: TPrim ts '[] Type

Natural :: TPrim ts '[] Type

Record :: AggType (DKind ts Kind) Type ls as -> TPrim ts as Type

List :: TPrim ts '[] (Type :~> Type)

Optional :: TPrim ts '[] (Type :~> Type)

TRecordLit :: RecordVal DSort (DKind ts) Kind ls ks at bs as -> TPrim ts as (KP (KRecord at) bs)

TUnionLit :: UnionVal DSort (DKind ts) Kind ls ks at bs a -> TPrim ts '[a] (KP (KRecord at) bs)

data DType ts :: [DKind ts Kind] -> DKind ts Kind -> Type where

Represents the possible types encountered in Dhall. A value of type

DTerm '[r, s] '[k, j] a

Describes a type of kind a with possible:

Kind variables of sort r, s

Type variables of kind k, j

Something of type DType '[] '[] a is a type of kind a with no free variables.

Constructors

TVar :: Index us a -> DType ts us a

TLam :: SDKind ts Kind u -> DType ts (u ': us) a -> DType ts us (u :~> a)

TApp :: DType ts us (a :~> b) -> DType ts us a -> DType ts us b infixl 9

TP :: TPrim ts as a -> Prod (DType ts us) as -> DType ts us a

TPoly :: SDSort t -> DType (t ': ts) (Map (KShiftSym ts (t ': ts) t Kind InsZ) us) a -> DType ts us (KPi (SDSortOf t) a)

TInst :: DType ts us (KPi (SDSortOf t) b) -> SDKind ts t a -> DType ts us (KSub (t ': ts) ts t Kind DelZ a b)

(:->) :: DType ts us Type -> DType ts us Type -> DType ts us Type infixr 0

Pi :: SDKind ts Kind u -> DType ts (u ': us) a -> DType ts us a

data SomeType ts :: [DKind ts Kind] -> Type where

Constructors

SomeType :: SDKind ts Kind a -> DType ts us a -> SomeType ts us

type Type = Type

The kind of types with values. For example Int :: Type.

type TBool = TP Bool Ø

type TNatural = TP Natural Ø

type TList = TP List Ø

type TOptional = TP Optional Ø

type (:$) = TApp infixl 9

type (:->) a b = a :-> b infixr 0

type family Shift ts us qs a b (ins :: Insert us qs a) (x :: DType ts us b) :: DType ts qs b where ...

Shift all type variables in a type expression of kind b to account for a new bound variable of kind a, to be inserted at the position indicated by the Insert.

Terms

data Prim ts us :: [DType ts us Type] -> DType ts us Type -> Type where

Primitives of Dhall terms, built into the language.

Constructors

BoolLit :: Bool -> Prim ts us '[] TBool

NaturalLit :: Natural -> Prim ts us '[] TNatural

NaturalFold :: Prim ts us '[] (TNatural :-> Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)))

NaturalBuild :: Prim ts us '[] (Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)) :-> TNatural)

NaturalPlus :: Prim ts us '[TNatural, TNatural] TNatural

NaturalTimes :: Prim ts us '[TNatural, TNatural] TNatural

NaturalIsZero :: Prim ts us '[] (TNatural :-> TBool)

ListFold :: Prim ts us '[] (Pi SType ((TList :$ TVar IZ) :-> Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ))))

ListBuild :: Prim ts us '[] (Pi SType (Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)) :-> (TList :$ TVar IZ)))

ListAppend :: Prim ts us '[TList :$ a, TList :$ a] (TList :$ a)

ListHead :: Prim ts us '[] (Pi SType ((TList :$ TVar IZ) :-> (TOptional :$ TVar IZ)))

ListLast :: Prim ts us '[] (Pi SType ((TList :$ TVar IZ) :-> (TOptional :$ TVar IZ)))

ListReverse :: Prim ts us '[] (Pi SType ((TList :$ TVar IZ) :-> (TList :$ TVar IZ)))

Some :: Prim ts us '[a] (TOptional :$ a)

None :: Prim ts us '[] (Pi SType (TOptional :$ TVar IZ))

RecordLit :: RecordVal (DKind ts Kind) (DType ts us) Type ls ks at bs as -> Prim ts us as (TP (Record at) bs)

UnionLit :: UnionVal (DKind ts Kind) (DType ts us) Type ls ks at bs a -> Prim ts us '[a] (TP (Record at) bs)

data PrimF ts us :: (Type -> Type) -> DType ts us (Type :~> Type) -> Type where

Primitive functors of Dhall terms, built into the language.

Constructors

ListLit :: PrimF ts us Seq TList

OptionalLit :: PrimF ts us Maybe TOptional

data DTerm ts (us :: [DKind ts Kind]) :: [DType ts us Type] -> DType ts us Type -> Type where

Represents the possible terms encountered in Dhall. A value of type

DTerm '[r, s] '[k, j] '[a, b, c] d

Describes a term of type d with possible:

Kind variables of sort r, s

Type variables of kind k, j

Variables of type a, b, c

Something of type DTerm '[] '[] '[] a is a term of type a with no free variables.

Constructors

Var :: Index vs a -> DTerm ts us vs a

Lam :: SDType ts us Type v -> DTerm ts us (v ': vs) a -> DTerm ts us vs (v :-> a)

App :: DTerm ts us vs (a :-> b) -> DTerm ts us vs a -> DTerm ts us vs b

P :: Prim ts us as a -> Prod (DTerm ts us vs) as -> DTerm ts us vs a

PF :: PrimF ts us f g -> SDType ts us Type a -> f (DTerm ts us vs a) -> DTerm ts us vs (g :$ a)

Poly :: SDKind ts Kind u -> DTerm ts (u ': us) (Map (ShiftSym ts us (u ': us) u Type InsZ) vs) a -> DTerm ts us vs (Pi (SDKindOf ts Kind u) a)

Inst :: DTerm ts us vs (Pi (SDKindOf ts Kind u) b) -> SDType ts us u a -> DTerm ts us vs (Sub ts (u ': us) us u Type DelZ a b)

data SomeTerm ts us :: [DType ts us Type] -> Type where

Constructors

SomeTerm :: SDType ts us Type a -> DTerm ts us vs a -> SomeTerm ts us vs

Mixed

data DExpr ts us :: [DType ts us Type] -> Fin N5 -> Type where

A DExpr fully covers all legal type-checking dhall terms. A value of type

DExpr '[ r, s ] '[ k, j ] '[ a, b ] n

Represents a dhall expression on level n ('FZ = term, 'FS 'FZ = type, etc.) with potential:

Kind variables of sort r, s

Type variables of kind k, j

Term variables of type a, b

A value of type DExpr '[] '[] '[] n represents a typed dhall expression with no free variables.

You can pattern match on it to get a value of one of the "levels" of the dhall type hierarchy, and also to get the "type" and representation of it.

The number of level goes up to 4 :

F0: term

F1: type

F2: kind

F3: sort

F4: "meta" level, outside of the hierarchy. This is necessary because the original untyped Dhall AST itself contains this level.

Note that you can restrict this to only DExpr past a given "level" by asking for or returning a DExpr ts us vs ('FS n), for instance. Such a value will only contain types, kinds, sorts, or meta. A DExpr ts us vs ('FS ('FS n)) will only contain kinds, sorts, or meta, etc.

Constructors

DEMeta :: DExpr ts us vs F4

DESort :: DSort -> DExpr ts us vs F3

DEKind :: SomeKind ts -> DExpr ts us vs F2

DEType :: SomeType ts us -> DExpr ts us vs F1

DETerm :: SomeTerm ts us vs -> DExpr ts us vs F0

data SomeDExpr ts us :: [DType ts us Type] -> Type where

Hides the "level" of a DExpr. Pattern match to find it. Can be useful when returning a DExpr of level unknown until runtime, or storing DExpr of multiple levels in a container.

Constructors

SomeDExpr :: SFin N5 l -> DExpr ts us vs l -> SomeDExpr ts us vs

dExprType :: DExpr ts us vs n -> DExpr ts us vs (ShiftFin N5 n)

Get the meta-level "type" of a DExpr. If it's a term, this will return its type. If it's a type, this returns its type, etc. It essentially goes up one "level" of the Dhall type hierarchy.

This will not typecheck if given a "Level 4" fin, so you cannot pass in DEMeta.

Shared

data AggType k :: k -> [Symbol] -> [k] -> Type where

Meta-level type describing a collection or aggregation of types. Used for specifying records and unions.

Constructors

ATZ :: AggType k r '[] '[]

ATS :: AggType k r ls as -> AggType k r (l ': ls) (r ': as)

data RecordVal k (j :: k -> Type) (r :: k) (ls :: [Symbol]) (ks :: [k]) :: AggType k r ls ks -> Prod j ks -> [j r] -> Type where

GADT for specifying a record value matching an AggType.

Basically you just need to stack as many RVS RVZ as there are fields in the record.

Constructors

RVZ :: RecordVal k j r '[] '[] ATZ Ø '[]

RVS :: RecordVal k j r ls ks at bs as -> RecordVal k j r (l ': ls) (r ': ks) (ATS at) (b :< bs) (a ': as)

data UnionVal k (j :: k -> Type) (r :: k) (ls :: [Symbol]) (ks :: [k]) :: AggType k r ls ks -> Prod j ks -> j r -> Type where

GADT for specifying a union value matching an AggType.

Constructors

UnionVal :: SIndex ks r i -> UnionVal k j r ls ks at bs (IxProd j ks r bs i)

Singletons

data SDSort :: DSort -> Type where

Constructors

SKind :: SDSort Kind

(:%*>) :: SDSort s -> SDSort t -> SDSort (s :*> t)

data SDKind ts a :: DKind ts a -> Type where

Constructors

SKVar :: SIndex ts a i -> SDKind ts a (KVar i)

SKLam :: SDSort t -> SDKind (t ': ts) a x -> SDKind ts (t :*> a) (KLam (SDSortOf t) x)

SKApp :: SDKind ts (a :*> b) f -> SDKind ts a x -> SDKind ts b (KApp f x)

(:%~>) :: SDKind ts Kind x -> SDKind ts Kind y -> SDKind ts Kind (x :~> y)

SType :: SDKind ts Kind Type

data STPrim ts as a :: TPrim ts as a -> Type where

Constructors

SBool :: STPrim ts '[] Type Bool

SNatural :: STPrim ts '[] Type Natural

SList :: STPrim ts '[] (Type :~> Type) List

SOptional :: STPrim ts '[] (Type :~> Type) Optional

data SDType ts us a :: DType ts us a -> Type where

Constructors

STVar :: SIndex us a i -> SDType ts us a (TVar i)

STLam :: SDKind ts Kind u -> SDType ts (u ': us) a x -> SDType ts us (u :~> a) (TLam (SDKindOf ts Kind u) x)

STApp :: SDType ts us (a :~> b) f -> SDType ts us a x -> SDType ts us b (TApp f x)

(:%->) :: SDType ts us Type x -> SDType ts us Type y -> SDType ts us Type (x :-> y)

STP :: STPrim ts as a x -> SProd (DType ts us) as p -> SDType ts us a (TP x p)

data SPrim ts us as a :: Prim ts us as a -> Type where

Constructors

SBoolLit :: Sing b -> SPrim ts us '[] TBool (BoolLit b)

SNaturalLit :: Sing n -> SPrim ts us '[] TNatural (NaturalLit n)

data SPrimF ts us f g :: PrimF ts us f g -> Type where

Constructors

SListLit :: SPrimF ts us Seq TList ListLit

SOptionalLit :: SPrimF ts us Maybe TOptional OptionalLit

data SDTerm ts us vs a :: DTerm ts us vs a -> Type where

Constructors

SVar :: SIndex vs a i -> SDTerm ts us vs a (Var i)

data KShiftSym ts ps a b :: Insert ts ps a -> DKind ts b ~> DKind ps b

data ShiftSym ts us qs a b :: Insert us qs a -> DType ts us b ~> DType ts qs b

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Util

type family Map (f :: a ~> b) (xs :: [a]) :: [b] where ...

Equations

Map f '[] = '[]

Map f (x ': xs) = (f @@ x) ': Map f xs

data MapSym (f :: a ~> b) :: [a] ~> [b]