Dhall.Typed.LC

Description

Implementing a typed lambda calculus. Uses De Bruijn indices in separate namespaces (that is, kind, type, term variables all have their own separate De Bruijn indexing)

We have a hierarchy: Term -> Type -> Kind -> Sort. We can stop at Sort because we don't allow any Sort variables, so there is no need to talk about the "type" of any Sorts. Disallowing variables effectively stops the chain.

It looks like each level has some sets of constructors just from its position on the chain.

At all levels, we have primitives. Term-level primitives are built-in functions and values and literals. Type-level primitives are built-in types and type functions. Dhall has no kind- or sort-level primitives.

If a level has variables (Terms, Types, Kinds), we gain some constructors:

Var, the variable constructor j Lam, the function abstraction

App, the function application

If a level has variables and its "type" also has variables (Terms, Types), we gain some constructors:

Poly, the type abstraction

Inst (instantiation), the type application

If a level is the "type" of something (n >= 1), we gain a function type constructor, the type of function abstractions in the immediately lower level.

If a level is the "type" of a "type" of something (n >= 2), we gain an "constant" constructor that links the two layers below it together. Examples are the Kind sort that links Kinds to Types (all types have kinds of sort Kind), and the Type kind that links types to values (all terms have types of kind Type)

If a type is the "type" of something (n >= 1), and it has variables, we have a Pi type constructor, the type of type abstractions (type-polymorphic values) in the lower level.

We actually gain a Pi type for our current level, and each level above our current level (that has variables). So for the Type level, we have Type-Pi and Kind-Pi, and for the Kind level, we have Kind-Pi.

So to summarize:

n = 3: Sort has a constant (it is n >= 2) and a function type constructor (it is n >= 1). It could also potentially have primitives, but Dhall has no sort primitives.

n = 2: Kind has variables, so it has Var, Lam, and App. It has a constant (they are n >= 2) and a function type constructor and a Pi constructor (it is n >= 1 and has variables). It could also potentially have primitives; Dhall has only record and union kinds as primitives.

n = 1: Type has variables, so it has Var, Lam, and App. It has a function type constructor and a Pi constructor (it is n >= 1 and has variables). Because its "type" has variables (Kind), it also has Poly and Inst. It has primitives, and in Dhall ther are several.

n = 0: Term has variables, so it has Var, Lam, and App. Because its "type" has variables, it has Poly and Inst. In Dhall, it has multitudes of primitives.