Dhall.Typed.Type.Index

Index

data Index (a :: [k]) (b :: k) :: forall k. [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: forall k (a :: [k]) (b :: k) (as :: [k]). Index (b ': as) b

IS :: forall k (a :: [k]) (b :: k) (bs :: [k]) (b1 :: k). Index bs b -> Index (b1 ': bs) b

data SIndex (as :: [k]) (a :: k) (b :: Index as a) :: forall k (as :: [k]) (a :: k). Index as a -> Type where

Kind-indexed singleton for Index. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SIndex', which has an actual proper Sing instance.

Since: decidable-0.1.5.0

Constructors

SIZ :: forall k (as :: [k]) (a :: k) (b :: Index as a) (as1 :: [k]). SIndex (a ': as1) a (IZ :: Index (a ': as1) a)

SIS :: forall k (as :: [k]) (a :: k) (b :: Index as a) (bs :: [k]) (i :: Index bs a) (b1 :: k). SIndex bs a i -> SIndex (b1 ': bs) a (IS i :: Index (b1 ': bs) a)

indexN :: Index as a -> N

Delete

data Delete (a :: [k]) (b :: [k]) (c :: k) :: forall k. [k] -> [k] -> k -> Type where

A Delete as bs x is a witness that you can delete item x from as to produce the list bs. It is essentially Insert flipped.

Some examples:

DelZ             :: Delete '[1,2,3] '[2,3] 1
DelS DelZ        :: Delete '[1,2,3] '[2,3] 2
DelS (DelS DelZ) :: Delete '[1,2,3] '[1,2] 3

bs will always be exactly one item shorter than as.

Constructors

DelZ :: forall k (a :: [k]) (b :: [k]) (c :: k). Delete (c ': b) b c

DelS :: forall k (a :: [k]) (b :: [k]) (c :: k) (as :: [k]) (bs :: [k]) (a1 :: k). Delete as bs c -> Delete (a1 ': as) (a1 ': bs) c

Insert

data Insert (a :: [k]) (b :: [k]) (c :: k) :: forall k. [k] -> [k] -> k -> Type where

An Insert as bs x is a witness that you can insert x into some position in list as to produce list bs. It is essentially Delete flipped.

Some examples:

InsZ                   :: Insert '[1,2,3] '[4,1,2,3] 4
InsS InsZ              :: Insert '[1,2,3] '[1,4,2,3] 4
InsS (InsS InsZ)       :: Insert '[1,2,3] '[1,2,4,3] 4
InsS (InsS (InsS InsZ) :: Insert '[1,2,3] '[1,2,3,4] 4

bs will always be exactly one item longer than as.

Constructors

InsZ :: forall k (a :: [k]) (b :: [k]) (c :: k). Insert a (c ': a) c

InsS :: forall k (a :: [k]) (b :: [k]) (c :: k) (as :: [k]) (bs :: [k]) (a1 :: k). Insert as bs c -> Insert (a1 ': as) (a1 ': bs) c

Orphan instances

PolySingKind (Index as a)

Instance details

Methods

fromPolySing :: PolySing (Index as a) x -> Index as a

toPolySing :: Index as a -> SomePolySing (Index as a)

SingEq (Index as a) (Index as b)

Instance details

Methods

singEq :: PolySing (Index as a) x -> PolySing (Index as b) y -> Decision (x :~~: y)

PolySingI (IZ :: Index (b ': as) b)

Instance details

Methods

polySing :: PolySing k IZ

PolySingI x => PolySingI (IS x :: Index (b1 ': bs) b)

Instance details

Methods

polySing :: PolySing k (IS x)

PolySingKind (Insert a b c)

Instance details

Methods

fromPolySing :: PolySing (Insert a b c) x -> Insert a b c

toPolySing :: Insert a b c -> SomePolySing (Insert a b c)

PolySingKind (Delete a b c)

Instance details

Methods

fromPolySing :: PolySing (Delete a b c) x -> Delete a b c

toPolySing :: Delete a b c -> SomePolySing (Delete a b c)

PolySingI (InsZ :: Insert a (c ': a) c)

Instance details

Methods

polySing :: PolySing k InsZ

PolySingI (DelZ :: Delete (c ': b) b c)

Instance details

Methods

polySing :: PolySing k DelZ

SingEq (Insert a1 a2 a3) (Insert a1 a2 b)

Instance details

Methods

singEq :: PolySing (Insert a1 a2 a3) x -> PolySing (Insert a1 a2 b) y -> Decision (x :~~: y)

SingEq (Delete a1 a2 a3) (Delete a1 b1 b2)

Instance details

Methods

singEq :: PolySing (Delete a1 a2 a3) x -> PolySing (Delete a1 b1 b2) y -> Decision (x :~~: y)

PolySingI x => PolySingI (InsS x :: Insert (a1 ': as) (a1 ': bs) c)

Instance details

Methods

polySing :: PolySing k (InsS x)

PolySingI x => PolySingI (DelS x :: Delete (a1 ': as) (a1 ': bs) c)

Instance details

Methods

polySing :: PolySing k (DelS x)