Dhall.Typed.Core

Kinds

data DKind

Represents the possible kinds encountered in Dhall.

Constructors

Type

DKind :~> DKind

data SDKind :: DKind -> Type where

Singletons for DKind. These are defined independently of Sing to avoid limitations of data family instances.

Note that at the moment, kind variables are not yet supported.

Constructors

SType :: SDKind Type

(:%~>) :: SDKind a -> SDKind b -> SDKind (a :~> b)

class SDKindI (k :: DKind) where

Typeclass for automatically generating singletons for a DType. Analogous to SingI for Sing.

Methods

sdKind :: SDKind k

sameDKind :: SDKind k -> SDKind j -> Maybe (k :~: j)

Compare two type-level DKind for equality.

Types

data DType :: [DKind] -> DKind -> Type where

Represents the possible types encountered in Dhall. A value of type

DType '[k1, k2, k3] k

Describes a type of kind k possibly containing free type variables of kind k1, k2, and k3.

Something of type DType '[] k is a type of kind k with no free variables.

Constructors

TVar :: Index us a -> DType us a

Pi :: SDKind a -> DType (a ': us) b -> DType us b

(:$) :: DType us (a :~> b) -> DType us a -> DType us b infixl 9

(:->) :: DType us Type -> DType us Type -> DType us Type infixr 0

TLet :: Bindings DKind DType vs us -> DType us a -> DType vs a

Bool :: DType us Type

Natural :: DType us Type

List :: DType us (Type :~> Type)

Optional :: DType us (Type :~> Type)

data SomeType :: [DKind] -> Type where

Constructors

SomeType :: SDKind k -> DType us k -> SomeType us

data SDType us k :: DType us k -> Type where

Singletons for DType. These are defined independently of Sing mostly to move the kind parameters to the front, to make them more easy to use.

Constructors

STVar :: SIndex us a i -> SDType us a (TVar i)

SPi :: SDKind a -> SDType (a ': us) b x -> SDType us b (Pi (SDKindOf a) x)

(:%$) :: SDType us (a :~> b) f -> SDType us a x -> SDType us b (f :$ x) infixl 9

(:%->) :: SDType us Type x -> SDType us Type y -> SDType us Type (x :-> y) infixr 0

SBool :: SDType us Type Bool

SNatural :: SDType us Type Natural

SList :: SDType us (Type :~> Type) List

SOptional :: SDType us (Type :~> Type) Optional

class SDTypeI us k (a :: DType us k) where

Typeclass for automatically generating singletons for a DType. Analogous to SingI for Sing, but with explicit kind parameters.

Methods

sdType :: SDType us k a

sameDType :: SDType '[] k a -> SDType '[] k b -> Maybe (a :~: b)

Compare two type-level DType with no free variables for equality.

sameDTypeWith :: Prod SDKind us -> SDType us k a -> SDType us k b -> Maybe (a :~: b)

Compare two type-level DType with free variables for equality by providing the kinds of each of the free variables.

kindOf :: SDType '[] k t -> SDKind k

Find the kind of a type singleton with no free variables.

kindOfWith :: Prod SDKind us -> SDType us k t -> SDKind k

Find the kind of a type singleton with free variables by providing the kinds of each free variable.

type family Sub as bs a b (d :: Delete as bs a) (x :: DType bs a) (r :: DType as b) :: DType bs b where ...

Substitute in a value for a given variable.

Equations

Sub as bs a b d x (TVar i) = MaybeVar a b x (Del as bs a b d i)

Sub as bs a b d x (Pi (u :: SDKind k) e) = Pi u (Sub (k ': as) (k ': bs) a b (DS d) (Shift bs (k ': bs) k a InsZ x) e)

Sub as bs a b d x ((i :: DType as (k :~> b)) :$ (j :: DType as k)) = Sub as bs a (k :~> b) d x i :$ Sub as bs a k d x j

Sub as bs a Type d x (i :-> j) = Sub as bs a Type d x i :-> Sub as bs a Type d x j

Sub as bs a Type d x Bool = Bool

Sub as bs a Type d x Natural = Natural

Sub as bs a (Type :~> Type) d x List = List

Sub as bs a (Type :~> Type) d x Optional = Optional

Sub as bs a b d x r = TypeError (Text "No Sub: " :<>: ShowType '(as, bs, a, b, d, x, r))

type family Shift as bs a b (ins :: Insert as bs a) (x :: DType as b) :: DType bs b where ...

Shift all variables to accomodate for a new bound variable.

Equations

Shift as bs a b ins (TVar i) = TVar (Ins as bs a b ins i)

Shift as bs a b ins (Pi (u :: SDKind k) e) = Pi u (Shift (k ': as) (k ': bs) a b (InsS ins) e)

Shift as bs a r i ((u :: DType as (k :~> r)) :$ (v :: DType as k)) = Shift as bs a (k :~> r) i u :$ Shift as bs a k i v

Shift as bs a Type i (u :-> v) = Shift as bs a Type i u :-> Shift as bs a Type i v

Shift as bs a Type i Bool = Bool

Shift as bs a Type i Natural = Natural

Shift as bs a (Type :~> Type) i List = List

Shift as bs a (Type :~> Type) i Optional = Optional

Shift as bs a b ins x = TypeError (Text "No Shift: " :<>: ShowType '(as, bs, a, b, ins, x))

type family MapShift (k :: DKind) (us :: [DKind]) (vs :: [DType us Type]) :: [DType (k ': us) Type] where ...

Equations

MapShift k us '[] = '[]

MapShift k us (v ': vs) = Shift us (k ': us) k Type InsZ v ': MapShift k us vs

Terms

data DTerm (us :: [DKind]) :: [DType us Type] -> DType us Type -> Type where

Represents the possible terms encountered in Dhall. A value of type

DTerm '[a, b, c] d

Describes a value of type d possibly containing free variables of type a, b, and c.

Something of type DTerm '[] a is a term of type a with no free variables.

Constructors

Var :: Index vs a -> DTerm us vs a

Lam :: SDType us Type a -> DTerm us (a ': vs) b -> DTerm us vs (a :-> b)

App :: DTerm us vs (a :-> b) -> DTerm us vs a -> DTerm us vs b infixl 3

Let :: Bindings (DType us Type) (DTerm us) vs qs -> DTerm us qs a -> DTerm us vs a

TLam :: SDKind k -> DTerm (k ': us) (MapShift k us vs) b -> DTerm us vs (Pi (SDKindOf k) b)

TApp :: DTerm us vs (Pi (SDKindOf k) b) -> SDType us k a -> DTerm us vs (Sub (k ': us) us k Type DZ a b) infixl 3

BoolLit :: Bool -> DTerm us vs Bool

NaturalLit :: Natural -> DTerm us vs Natural

NaturalFold :: DTerm us vs (Natural :-> Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)))

NaturalBuild :: DTerm us vs (Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)) :-> Natural)

NaturalPlus :: DTerm us vs Natural -> DTerm us vs Natural -> DTerm us vs Natural

NaturalTimes :: DTerm us vs Natural -> DTerm us vs Natural -> DTerm us vs Natural

NaturalIsZero :: DTerm us vs (Natural :-> Bool)

ListLit :: SDType us Type a -> Seq (DTerm us vs a) -> DTerm us vs (List :$ a)

ListFold :: DTerm us vs (Pi SType ((List :$ TVar IZ) :-> Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ))))

ListBuild :: DTerm us vs (Pi SType (Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)) :-> (List :$ TVar IZ)))

ListAppend :: DTerm us vs (List :$ a) -> DTerm us vs (List :$ a) -> DTerm us vs (List :$ a)

ListHead :: DTerm us vs (Pi SType ((List :$ TVar IZ) :-> (Optional :$ TVar IZ)))

ListLast :: DTerm us vs (Pi SType ((List :$ TVar IZ) :-> (Optional :$ TVar IZ)))

ListReverse :: DTerm us vs (Pi SType ((List :$ TVar IZ) :-> (List :$ TVar IZ)))

OptionalLit :: SDType us Type a -> Maybe (DTerm us vs a) -> DTerm us vs (Optional :$ a)

Some :: DTerm us vs a -> DTerm us vs (Optional :$ a)

None :: DTerm us vs (Pi SType (Optional :$ TVar IZ))

data Bindings k :: ([k] -> k -> Type) -> [k] -> [k] -> Type where

A non-empty series of Let bindings.

Constructors

BNil :: f vs a -> Bindings k f vs (a ': vs)

(:<?) :: f vs a -> Bindings k f (a ': vs) us -> Bindings k f vs us infixr 5

data SomeTerm us :: [DType us Type] -> Type where

Constructors

SomeTerm :: SDType us Type a -> DTerm us vs a -> SomeTerm us vs

data SDTerm us (vs :: [DType us Type]) a :: DTerm us vs a -> Type where

Singletons for DTerm. These are defined independently of Sing mostly to move the kind parameters to the front, to make them more easy to use.

Note that there is currently no singleton implemented for the TLam constructor.

Constructors

SVar :: SIndex vs a i -> SDTerm us vs a (Var i)

SLam :: SDType us Type a -> SDTerm us (a ': vs) b x -> SDTerm us vs (a :-> b) (Lam (SDTypeOf us Type a) x)

SApp :: SDTerm us vs (a :-> b) f -> SDTerm us vs a x -> SDTerm us vs b (App f x)

STLam :: SDKind k -> SDTerm (k ': us) (MapShift k us vs) b x -> SDTerm us vs (Pi (SDKindOf k) b) (TLam (SDKindOf k) x)

STApp :: SDTerm us vs (Pi (SDKindOf k) b) f -> SDType us k a -> SDTerm us vs (Sub (k ': us) us k Type DZ a b) (TApp f (SDTypeOf us k a))

SBoolLit :: Sing b -> SDTerm us vs Bool (BoolLit b)

SNaturalLit :: Sing n -> SDTerm us vs Natural (NaturalLit n)

SNaturalFold :: SDTerm us vs (Natural :-> Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ))) NaturalFold

SNaturalBuild :: SDTerm us vs (Pi SType ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)) :-> Natural) NaturalBuild

SNaturalPlus :: SDTerm us vs Natural x -> SDTerm us vs Natural y -> SDTerm us vs Natural (NaturalPlus x y)

SNaturalTimes :: SDTerm us vs Natural x -> SDTerm us vs Natural y -> SDTerm us vs Natural (NaturalTimes x y)

SNaturalIsZero :: SDTerm us vs (Natural :-> Bool) NaturalIsZero

SListLit :: SeqListEq xs xs' -> SDType us Type a -> Prod (SDTerm us vs a) xs' -> SDTerm us vs (List :$ a) (ListLit (SDTypeOf us Type a) xs)

SListFold :: SDTerm us vs (Pi SType ((List :$ TVar IZ) :-> Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)))) ListFold

SListBuild :: SDTerm us vs (Pi SType (Pi SType ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)) :-> (List :$ TVar IZ))) ListBuild

SListAppend :: SDTerm us vs (List :$ a) xs -> SDTerm us vs (List :$ a) ys -> SDTerm us vs (List :$ a) (ListAppend xs ys)

SListHead :: SDTerm us vs (Pi SType ((List :$ TVar IZ) :-> (Optional :$ TVar IZ))) ListHead

SListLast :: SDTerm us vs (Pi SType ((List :$ TVar IZ) :-> (Optional :$ TVar IZ))) ListLast

SListReverse :: SDTerm us vs (Pi SType ((List :$ TVar IZ) :-> (List :$ TVar IZ))) ListReverse

SOptionalLit :: SDType us Type a -> Option (SDTerm us vs a) o -> SDTerm us vs (Optional :$ a) (OptionalLit (SDTypeOf us Type a) o)

SSome :: SDTerm us vs a x -> SDTerm us vs (Optional :$ a) (Some x)

SNone :: SDTerm us vs (Pi SType (Optional :$ TVar IZ)) None

data SeqListEq :: Seq a -> [a] -> Type where

Constructors

SeqListEq :: SeqListEq xs ys

typeOf :: SDTerm '[] '[] a x -> SDType '[] Type a

Find the type of a term singleton with no free variables.

typeOfWith :: Prod (SDType us Type) vs -> SDTerm us vs a x -> SDType us Type a

Find the type of a term singleton with free variables by providing the type of each free variable.

Expr

data SomeExpr us :: [DType us Type] -> Type where

Constructors

SEKind :: DKind -> SomeExpr us vs

SEType :: SomeType us -> SomeExpr us vs

SETerm :: SomeTerm us vs -> SomeExpr us vs

Evaluation

type family DTypeRep us (p :: Prod (DType us) us) k (x :: DType us k) :: DKindRep k where ...

Matches a DType to the actual Haskell type that it represents.

Equations

DTypeRep us p k (TVar i) = DTypeRep us p k (IxProd (DType us) us k p i)

DTypeRep us p Type (Pi (u :: SDKind a) t) = Forall us p a t

DTypeRep us p (k :~> Type) (Pi (u :: SDKind a) t) = ForallTC us p k a t

DTypeRep us p k ((f :: DType us (r :~> k)) :$ (x :: DType us r)) = DTypeRep us p (r :~> k) f (DTypeRep us p r x)

DTypeRep us p Type (a :-> b) = DTypeRep us p Type a -> DTypeRep us p Type b

DTypeRep us p Type Bool = Bool

DTypeRep us p Type Natural = Natural

DTypeRep us p (Type :~> Type) List = Seq

DTypeRep us p (Type :~> Type) Optional = Maybe

DTypeRep us p k x = TypeError (Text "No DTypeRep: " :<>: ShowType '(k, x))

data Forall us p k :: DType (k ': us) Type -> Type where

A value of a polymorphic type.

Constructors

FA

Fields

:: { runForall :: forall r. SDType us k r -> DTypeRep us p Type (Sub (k ': us) us k Type DZ r a)

} -> Forall us p k a

data ForallTC us p j k :: DType (k ': us) (j :~> Type) -> DKindRep j -> Type where

A value of a polymorphic type, lifted to take a type constructor as a parameter.

Constructors

FATC

Fields

:: { runForallTCC :: forall r. SDType us k r -> DTypeRep us p (j :~> Type) (Sub (k ': us) us k (j :~> Type) DZ r a) x

} -> ForallTC us p j k a x

newtype DTypeRepVal us p (a :: DType us Type)

Newtype wrapper over a Haskell value of the DTypeRep of that term.

Constructors

DTRV

Fields

getDTRV :: DTypeRep us p Type a

fromTerm :: DTerm '[] '[] a -> DTypeRep '[] Ø Type a

Turn a DTerm with no free variables into a Haskell value of the appropriate type.

fromTermWith :: forall vs a. Prod (DTypeRepVal '[] Ø) vs -> DTerm '[] vs a -> DTypeRep '[] Ø Type a

Turn a DTerm with free variables into a Haskell value of the appropriate type by providing values for each free variable.

toTerm :: SDType '[] Type a -> DTypeRep '[] Ø Type a -> Maybe (DTerm '[] '[] a)

Attempt to convert a Haskell value into a DTerm with no free variables. This will fail if you attempt to convert any Haskell functions a -> b, since we cannot encode these in general into a finite language like Dhall.

Manipulation

sShift :: SDType as k x -> SDType (a ': as) k (Shift as (a ': as) a k InsZ x)

Allows you to use a type variable "under" a TLam.

sShift_ :: SInsert as bs a ins -> SDType as b x -> SDType bs b (Shift as bs a b ins x)

Like sShift, but can shift a type variable under multiple TLams.

Providing SInsZ will shift a single layer, SInsS SInsZ will shift two layers, etc.

subIns :: forall k j a b. SDType '[] k a -> SDType '[] j b -> a :~: Sub '[j] '[] j k (DZ :: Delete '[j] '[] j) b (Shift '[] '[j] j k (InsZ :: Insert '[] '[j] j) a)

Required equality witness for using a type variable under a TLam.

This is automatically resolved if you turn on the typechecker plugin.

{--}

subIns2 :: SDType '[] k a -> SDType '[] j b -> SDType '[] l c -> a :~: Sub '[l] '[] l k DZ c (Sub '[l, j] '[l] j k (DS DZ) (Shift '[] '[l] l j InsZ b) (Shift '[j] '[l, j] l k InsZ (Shift '[] '[j] j k InsZ a)))

Like subIns, but for two layers of TLam.

This is automatically resolved if you turn on the typechecker plugin. The typechecker plugin will solve arbitrarily nested layers.

{--}

sSub :: SDType bs a x -> SDType (a ': bs) b r -> SDType bs b (Sub (a ': bs) bs a b DZ x r)

Substitute a type into the first free variable of a type expression.

sSub_ :: SDelete as bs c del -> SDType bs c x -> SDType as b r -> SDType bs b (Sub as bs c b del x r)

Substitute a type into the Nth free variable of a type expression. Providing DZ will substitute in the first free variable, providing DS DZ will substitute in the second free variable, etc.

shiftProd :: forall us vs k. Prod (SDType us Type) vs -> Prod (SDType (k ': us) Type) (MapShift k us vs)

Singletons

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.