Dhall.Typed.Core

Expression

Sorts

data DSort :: Type where

Represents the possible sorts encountered in Dhall.

Note that this implementation allows records of kinds with sorts other than Kind, so { foo : Kind -> Kind } would typecheck, even though normal Dhall forbids this.

A DSort is always in normal form.

Constructors

Kind :: DSort

(:*>) :: DSort -> DSort -> DSort

KRecord :: AggType () ls as -> DSort

KUnion :: AggType () ls as -> DSort

Kinds

data DKind :: [DSort] -> DSort -> Type where

Represents the possible types encountered in Dhall. A value of type

DKind '[r, s] t

Describes a kind of sort t with possible kind variables of sorts r and s.

Something of type DKind '[] t is a kind of sort t with no free variables.

Constructors

KVar :: Index ts a -> DKind ts a

KLam :: SDSort t -> DKind (t ': ts) a -> DKind ts (t :*> a)

KApp :: DKind ts (a :*> b) -> DKind ts a -> DKind ts b

(:~>) :: DKind ts Kind -> DKind ts Kind -> DKind ts Kind infixr 1

KPi :: SDSort t -> DKind (t ': ts) a -> DKind ts a

Type :: DKind ts Kind

data SomeKind :: [DSort] -> Type where

Constructors

SomeKind :: SDSort a -> DKind ts a -> SomeKind ts

type (:~>) a b = a :~> b infixr 1

type family KShift ts ps a b (ins :: Insert ts ps a) (x :: DKind ts b) :: DKind ps b where ...

Shift all kind variables in a kind expression of sort b to account for a new bound variable of sort a, to be inserted at the position indicated by the Insert.

toSomeKind :: PolySingI a => DKind ts a -> SomeKind ts

type family KNormalize ts a (x :: DKind ts a) :: DKind ts a where ...

Ideally we would want this to be encodable within the type. But the main problem here is checking if the LHS of an application is a variable or not. This is the next best thing?

Equations

KNormalize ts a (KVar i) = KVar i

KNormalize ts (t :*> a) (KLam tt x) = KLam tt (KNormalize (t ': ts) a x)

KNormalize ts a (KApp (KLam (tt :: SDSort t) f) x) = KNormalize ts a (KSub (t ': ts) ts t a DelZ x f)

KNormalize ts a (KApp (f :: DKind ts (r :*> a)) x) = KApp (KNormalize ts (r :*> a) f) (KNormalize ts r x)

KNormalize ts Kind (x :~> y) = KNormalize ts Kind x :~> KNormalize ts Kind y

KNormalize ts a (KPi (tt :: SDSort t) x) = KPi tt (KNormalize (t ': ts) a x)

KNormalize ts Kind Type = Type

KNormalize ts a t = TypeError (Text "No KNormalize")

data NDKind ts a :: DKind ts a -> Type where

Version of SDKind that exposes itself in normal form.

Constructors

NDK :: SDKind ts a x -> NDKind ts a (KNormalize ts a x)

type family KSub ts rs a b (del :: Delete ts rs a) (x :: DKind rs a) (r :: DKind ts b) :: DKind rs b where ...

Substitute in a kind for all occurrences of a kind variable of sort a indicated by the Delete within a kind of osrt b.

Types

data DType ts :: [DKind ts Kind] -> DKind ts Kind -> Type where

Represents the possible types encountered in Dhall. A value of type

DTerm '[r, s] '[k, j] a

Describes a type of kind a with possible:

Kind variables of sort r, s

Type variables of kind k, j

Something of type DType '[] '[] a is a type of kind a with no free variables.

Note that the type of "kind-polymorphic values" (functions from kinds to terms) is not yet supported.

The kinds of the type variables should all be normalized. All constructors that introduce type variables should normalize automatically, but it's possible create nonsensical types with TVar.

Constructors

TVar :: Index us a -> DType ts us a

TLam :: NDKind ts Kind u -> DType ts (u ': us) a -> DType ts us (u :~> a)

TApp :: DType ts us (a :~> b) -> DType ts us a -> DType ts us b infixl 9

TPoly :: SingSing DSort t (WS tt) -> DType (t ': ts) (Map (KShiftSym ts (t ': ts) t Kind InsZ) us) a -> DType ts us (KPi tt a)

TInst :: DType ts us (KPi tt b) -> SDKind ts t a -> DType ts us (KSub (t ': ts) ts t Kind DelZ a b)

(:->) :: DType ts us Type -> DType ts us Type -> DType ts us Type infixr 0

Pi :: NDKind ts Kind u -> DType ts (u ': us) a -> DType ts us a

Bool :: DType ts us Type

Natural :: DType ts us Type

List :: DType ts us (Type :~> Type)

Optional :: DType ts us (Type :~> Type)

data SomeType ts :: [DKind ts Kind] -> Type where

Constructors

SomeType :: NDKind ts Kind a -> DType ts us a -> SomeType ts us

type (:$) = TApp infixl 9

type (:->) a b = a :-> b infixr 0

type family Shift ts us qs a b (ins :: Insert us qs a) (x :: DType ts us b) :: DType ts qs b where ...

Shift all type variables in a type expression of kind b to account for a new bound variable of kind a, to be inserted at the position indicated by the Insert.

toSomeType :: forall ts us a. (PolySingI a, KNormalize ts Kind a ~ a) => DType ts us a -> SomeType ts us

type family TNormalize ts us a (x :: DType ts us a) :: DType ts us a where ...

Ideally we would want this to be encodable within the type. But the main problem here is checking if the LHS of an application is a variable or not. This is the next best thing?

Equations

TNormalize ts us a (TVar i) = TVar i

TNormalize ts us (u :~> a) (TLam (uu :: NDKind ts Kind u) x) = TLam uu (TNormalize ts (u ': us) a x)

TNormalize ts us a (TApp (f :: DType ts us (r :~> a)) x) = TApp (TNormalize ts us (r :~> a) f) (TNormalize ts us r x)

TNormalize ts us Type (x :-> y) = TNormalize ts us Type x :-> TNormalize ts us Type y

TNormalize ts us a (Pi (uu :: NDKind ts Kind u) x) = Pi uu (TNormalize ts (u ': us) a x)

TNormalize ts us Type Bool = Bool

TNormalize ts us Type Natural = Natural

TNormalize ts us (Type :~> Type) List = List

TNormalize ts us (Type :~> Type) Optional = Optional

TNormalize ts us a t = TypeError (Text "No TNormalize")

data NDType ts us a :: DType ts us a -> Type where

Version of SDType that exposes itself in normal form.

Constructors

NDT :: SDType ts us a x -> NDType ts us a (TNormalize ts us a x)

type family Sub ts us qs a b (del :: Delete us qs a) (x :: DType ts qs a) (r :: DType ts us b) :: DType ts qs b where ...

Substitute in a type for all occurrences of a type variable of kind a indicated by the Delete within a type of kind b.

normalizeKindOf :: DType ts us a -> DType ts us (KNormalize ts Kind a)

Terms

data Prim ts us :: [DType ts us Type] -> DType ts us Type -> Type where

Primitives of Dhall terms, built into the language.

Constructors

BoolLit :: Bool -> Prim ts us '[] Bool

NaturalLit :: Natural -> Prim ts us '[] Natural

NaturalFold :: Prim ts us '[] (Natural :-> Pi (NDK SType) ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)))

NaturalBuild :: Prim ts us '[] (Pi (NDK SType) ((TVar IZ :-> TVar IZ) :-> (TVar IZ :-> TVar IZ)) :-> Natural)

NaturalPlus :: Prim ts us '[Natural, Natural] Natural

NaturalTimes :: Prim ts us '[Natural, Natural] Natural

NaturalIsZero :: Prim ts us '[] (Natural :-> Bool)

ListFold :: Prim ts us '[] (Pi (NDK SType) ((List :$ TVar IZ) :-> Pi (NDK SType) ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ))))

ListBuild :: Prim ts us '[] (Pi (NDK SType) (Pi (NDK SType) ((TVar (IS IZ) :-> (TVar IZ :-> TVar IZ)) :-> (TVar IZ :-> TVar IZ)) :-> (List :$ TVar IZ)))

ListAppend :: SDType ts us Type a -> Prim ts us '[List :$ a, List :$ a] (List :$ a)

ListHead :: Prim ts us '[] (Pi (NDK SType) ((List :$ TVar IZ) :-> (Optional :$ TVar IZ)))

ListLast :: Prim ts us '[] (Pi (NDK SType) ((List :$ TVar IZ) :-> (Optional :$ TVar IZ)))

ListReverse :: Prim ts us '[] (Pi (NDK SType) ((List :$ TVar IZ) :-> (List :$ TVar IZ)))

Some :: SDType ts us Type a -> Prim ts us '[a] (Optional :$ a)

None :: Prim ts us '[] (Pi (NDK SType) (Optional :$ TVar IZ))

data DTerm ts (us :: [DKind ts Kind]) :: [DType ts us Type] -> DType ts us Type -> Type where

Represents the possible terms encountered in Dhall. A value of type

DTerm '[r, s] '[k, j] '[a, b, c] d

Describes a term of type d with possible:

Kind variables of sort r, s

Type variables of kind k, j

Variables of type a, b, c

Something of type DTerm '[] '[] '[] a is a term of type a with no free variables.

Note that "kind-polymorphic values" (functions from kinds to terms) are not yet supported.

Constructors

Var :: Index vs a -> DTerm ts us vs a

Lam :: NDType ts us Type v -> DTerm ts us (v ': vs) a -> DTerm ts us vs (v :-> a)

App :: DTerm ts us vs (a :-> b) -> DTerm ts us vs a -> DTerm ts us vs b

Poly :: SNDKind ts Kind u uu -> DTerm ts (u ': us) (Map (ShiftSym ts us (u ': us) u Type InsZ) vs) a -> DTerm ts us vs (Pi uu a)

Inst :: SNDKind ts Kind u uu -> DTerm ts us vs (Pi uu b) -> NDType ts us u a -> DTerm ts us vs (Sub ts (u ': us) us u Type DelZ a b)

P :: Prim ts us as a -> Prod (DTerm ts us vs) as -> DTerm ts us vs a

ListLit :: NDType ts us Type a -> [DTerm ts us vs a] -> DTerm ts us vs (List :$ a)

OptionalLit :: NDType ts us Type a -> Maybe (DTerm ts us vs a) -> DTerm ts us vs (Optional :$ a)

data SomeTerm ts us :: [DType ts us Type] -> Type where

Constructors

SomeTerm :: NDType ts us Type a -> DTerm ts us vs a -> SomeTerm ts us vs

toSomeTerm :: forall ts us vs a. (PolySingI a, TNormalize ts us Type a ~ a) => DTerm ts us vs a -> SomeTerm ts us vs

normalizeTypeOf :: DTerm ts us vs a -> DTerm ts us vs (TNormalize ts us Type a)

Mixed

data DExpr ts us :: [DType ts us Type] -> Fin N5 -> Type where

A DExpr fully covers all legal type-checking dhall terms. A value of type

DExpr '[ r, s ] '[ k, j ] '[ a, b ] n

Represents a dhall expression on level n ('FZ = term, 'FS 'FZ = type, etc.) with potential:

Kind variables of sort r, s

Type variables of kind k, j

Term variables of type a, b

A value of type DExpr '[] '[] '[] n represents a typed dhall expression with no free variables.

You can pattern match on it to get a value of one of the "levels" of the dhall type hierarchy, and also to get the "type" and representation of it.

The number of level goes up to 4 :

F0: term

F1: type

F2: kind

F3: sort

F4: order

Note that you can restrict this to only DExpr past a given "level" by asking for or returning a DExpr ts us vs ('FS n), for instance. Such a value will only contain types, kinds, sorts, or order A DExpr ts us vs ('FS ('FS n)) will only contain kinds, sorts, or order, etc.

Constructors

DEOrder :: DExpr ts us vs F4

DESort :: DSort -> DExpr ts us vs F3

DEKind :: SomeKind ts -> DExpr ts us vs F2

DEType :: SomeType ts us -> DExpr ts us vs F1

DETerm :: SomeTerm ts us vs -> DExpr ts us vs F0

data SomeDExpr ts us :: [DType ts us Type] -> Type where

Hides the "level" of a DExpr. Pattern match to find it. Can be useful when returning a DExpr of level unknown until runtime, or storing DExpr of multiple levels in a container.

Constructors

SomeDExpr :: DExpr ts us vs l -> SomeDExpr ts us vs

dExprType :: DExpr ts us vs n -> DExpr ts us vs (ShiftFin N5 n)

Get the meta-level "type" of a DExpr. If it's a term, this will return its type. If it's a type, this returns its type, etc. It essentially goes up one "level" of the Dhall type hierarchy.

This will not typecheck if given a "Level 4" fin, so you cannot pass in DEOrder.

deKind :: PolySingI a => DKind ts a -> DExpr ts us vs F2

deType :: (PolySingI a, KNormalize ts Kind a ~ a) => DType ts us a -> DExpr ts us vs F1

deTerm :: (PolySingI a, TNormalize ts us Type a ~ a) => DTerm ts us vs a -> DExpr ts us vs F0

Shared

data AggType k (ls :: [Text]) (as :: [k]) where

Meta-level type describing a collection or aggregation of types. Used for specifying records and unions. data AggType k :: [Text] -> [k] -> Type where

Constructors

ATZ :: AggType k '[] '[]

ATS :: SText l -> WrappedSing k (a :: k) -> AggType k ls as -> AggType k (l ': ls) (a ': as)

Type manipulation

sortOf :: DKind '[] a -> SDSort a

kindOf :: DType ts '[] a -> SDKind ts Kind a

typeOf :: DTerm ts us '[] a -> SDType ts us Type a

sortOfWith :: Prod SDSort ts -> DKind ts a -> SDSort a

kindOfWith :: Prod (SDKind ts Kind) us -> DType ts us a -> SDKind ts Kind a

typeOfWith :: Prod (SDType ts us Type) vs -> DTerm ts us vs a -> SDType ts us Type a

Singletons

Types

data SDSort (x :: DSort) where

Constructors

SKind :: SDSort Kind

(:%*>) :: forall x x. (SDSort x) -> (SDSort x) -> SDSort ((:*>) x x)

SKRecord :: forall (ls :: [Text]) (as :: [()]) x. (PolySing (AggType () ls as) x) -> SDSort (KRecord x)

SKUnion :: forall (ls :: [Text]) (as :: [()]) x. (PolySing (AggType () ls as) x) -> SDSort (KUnion x)

data SDKind (a :: [DSort]) (b :: DSort) (x :: DKind (a :: [DSort]) (b :: DSort)) where

Constructors

SKVar :: forall (ts :: [DSort]) (a :: DSort) x. (PolySing (Index ts a) x) -> SDKind ts a (KVar x)

SKLam :: forall (t :: DSort) (ts :: [DSort]) (a :: DSort) x x. (SingSing DSort t (WS x)) -> (SDKind ((:) t ts) a x) -> SDKind ts ((:*>) t a) (KLam x x)

SKApp :: forall (ts :: [DSort]) (a :: DSort) (b :: DSort) x x. (SDKind ts ((:*>) a b) x) -> (SDKind ts a x) -> SDKind ts b (KApp x x)

(:%~>) :: forall (ts :: [DSort]) x x. (SDKind ts Kind x) -> (SDKind ts Kind x) -> SDKind ts Kind ((:~>) x x)

SKPi :: forall (t :: DSort) (ts :: [DSort]) (a :: DSort) x x. (SingSing DSort t (WS x)) -> (SDKind ((:) t ts) a x) -> SDKind ts a (KPi x x)

SType :: forall (ts :: [DSort]). SDKind ts Kind Type

data SNDKind (ts :: [DSort]) (a :: DSort) (b :: DKind ts a) (x :: NDKind (ts :: [DSort]) (a :: DSort) (b :: DKind ts a)) where

Constructors

SNDK :: forall (ts :: [DSort]) (a :: DSort) (x :: DKind ts a) x. (SingSing (DKind ts a) x (WS x)) -> SNDKind ts a (KNormalize ts a x) (NDK x)

data SDType (ts :: [DSort]) (a :: [DKind ts Kind]) (b :: DKind ts Kind) (x :: DType (ts :: [DSort]) (a :: [DKind ts Kind]) (b :: DKind ts Kind)) where

Constructors

STVar :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ts Kind) x. (PolySing (Index us a) x) -> SDType ts us a (TVar x)

STLam :: forall (ts :: [DSort]) (u :: DKind ts Kind) (us :: [DKind ts Kind]) (a :: DKind ts Kind) x x. (PolySing (NDKind ts Kind u) x) -> (SDType ts ((:) u us) a x) -> SDType ts us ((:~>) u a) (TLam x x)

STApp :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ts Kind) (b :: DKind ts Kind) x x. (SDType ts us ((:~>) a b) x) -> (SDType ts us a x) -> SDType ts us b (TApp x x)

STPoly :: forall (t :: DSort) (tt :: SDSort t) (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ((:) t ts) Kind) x x. (SingSing (WrappedSing DSort t) (WS tt) (WS x)) -> (SDType ((:) t ts) (Map (KShiftSym ts ((:) t ts) t Kind (InsZ :: Insert ts ((:) t ts) t)) us) a x) -> SDType ts us (KPi tt a) (TPoly x x)

STInst :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (t :: DSort) (tt :: SDSort t) (b :: DKind ((:) t ts) Kind) (a :: DKind ts t) x x. (SDType ts us (KPi tt b) x) -> (SingSing (DKind ts t) a (WS x)) -> SDType ts us (KSub ((:) t ts) ts t Kind (DelZ :: Delete ((:) t ts) ts t) a b) (TInst x x)

(:%->) :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) x x. (SDType ts us (Type :: DKind ts Kind) x) -> (SDType ts us (Type :: DKind ts Kind) x) -> SDType ts us (Type :: DKind ts Kind) ((:->) x x)

SPi :: forall (ts :: [DSort]) (u :: DKind ts Kind) (us :: [DKind ts Kind]) (a :: DKind ts Kind) x x. (PolySing (NDKind ts Kind u) x) -> (SDType ts ((:) u us) a x) -> SDType ts us a (Pi x x)

SBool :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SDType ts us (Type :: DKind ts Kind) Bool

SNatural :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SDType ts us (Type :: DKind ts Kind) Natural

SList :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SDType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind)) List

SOptional :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SDType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind)) Optional

data SNDType (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ts Kind) (b :: DType ts us a) (x :: NDType (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ts Kind) (b :: DType ts us a)) where

Constructors

SNDT :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DKind ts Kind) (x :: DType ts us a) x. (SingSing (DType ts us a) x (WS x)) -> SNDType ts us a (TNormalize ts us a x) (NDT x)

data SPrim (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: [DType ts us (Type :: DKind ts Kind)]) (b :: DType ts us (Type :: DKind ts Kind)) (x :: Prim (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: [DType ts us (Type :: DKind ts Kind)]) (b :: DType ts us (Type :: DKind ts Kind))) where

Constructors

SBoolLit :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) x. (PolySing Bool x) -> SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Bool :: DType ts us (Type :: DKind ts Kind)) (BoolLit x)

SNaturalLit :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) x. (PolySing Natural x) -> SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Natural :: DType ts us (Type :: DKind ts Kind)) (NaturalLit x)

SNaturalFold :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) ((:->) (Natural :: DType ts us (Type :: DKind ts Kind)) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))))))) NaturalFold

SNaturalBuild :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) ((:->) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))))) (Natural :: DType ts us (Type :: DKind ts Kind))) NaturalBuild

SNaturalPlus :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ((:) (Natural :: DType ts us (Type :: DKind ts Kind)) ((:) (Natural :: DType ts us (Type :: DKind ts Kind)) ('[] :: [DType ts us (Type :: DKind ts Kind)]))) (Natural :: DType ts us (Type :: DKind ts Kind)) NaturalPlus

SNaturalTimes :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ((:) (Natural :: DType ts us (Type :: DKind ts Kind)) ((:) (Natural :: DType ts us (Type :: DKind ts Kind)) ('[] :: [DType ts us (Type :: DKind ts Kind)]))) (Natural :: DType ts us (Type :: DKind ts Kind)) NaturalTimes

SNaturalIsZero :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) ((:->) (Natural :: DType ts us (Type :: DKind ts Kind)) (Bool :: DType ts us (Type :: DKind ts Kind))) NaturalIsZero

SListFold :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:->) (TVar (IS (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)) :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind)))))))) ListFold

SListBuild :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:->) (TVar (IS (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)) :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))))) ((:->) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) ((:) (Type :: DKind ts Kind) us)) (Type :: DKind ts Kind)))))) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))))) ListBuild

SListAppend :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DType ts us (Type :: DKind ts Kind)) x. (SingSing (DType ts us (Type :: DKind ts Kind)) a (WS x)) -> SPrim ts us ((:) ((:$) (List :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) ((:) ((:$) (List :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) ('[] :: [DType ts us (Type :: DKind ts Kind)]))) ((:$) (List :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) (ListAppend x)

SListHead :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) ((:$) (Optional :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))))) ListHead

SListLast :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) ((:$) (Optional :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))))) ListLast

SListReverse :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:->) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))) ((:$) (List :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind)))))) ListReverse

SSome :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DType ts us (Type :: DKind ts Kind)) x. (SingSing (DType ts us (Type :: DKind ts Kind)) a (WS x)) -> SPrim ts us ((:) a ('[] :: [DType ts us (Type :: DKind ts Kind)])) ((:$) (Optional :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) (Some x)

SNone :: forall (ts :: [DSort]) (us :: [DKind ts Kind]). SPrim ts us ('[] :: [DType ts us (Type :: DKind ts Kind)]) (Pi (NDK (SType :: SDKind ts Kind (Type :: DKind ts Kind))) ((:$) (Optional :: DType ts ((:) (Type :: DKind ts Kind) us) ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) (TVar (IZ :: Index ((:) (Type :: DKind ts Kind) us) (Type :: DKind ts Kind))))) None

data SDTerm (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: [DType ts us (Type :: DKind ts Kind)]) (b :: DType ts us (Type :: DKind ts Kind)) (x :: DTerm (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: [DType ts us (Type :: DKind ts Kind)]) (b :: DType ts us (Type :: DKind ts Kind))) where

Constructors

SVar :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (vs :: [DType ts us (Type :: DKind ts Kind)]) (a :: DType ts us (Type :: DKind ts Kind)) x. (PolySing (Index vs a) x) -> SDTerm ts us vs a (Var x)

SLam :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (v :: DType ts us (Type :: DKind ts Kind)) (vs :: [DType ts us (Type :: DKind ts Kind)]) (a :: DType ts us (Type :: DKind ts Kind)) x x. (PolySing (NDType ts us (Type :: DKind ts Kind) v) x) -> (SDTerm ts us ((:) v vs) a x) -> SDTerm ts us vs ((:->) v a) (Lam x x)

SApp :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (vs :: [DType ts us (Type :: DKind ts Kind)]) (a :: DType ts us (Type :: DKind ts Kind)) (b :: DType ts us (Type :: DKind ts Kind)) x x. (SDTerm ts us vs ((:->) a b) x) -> (SDTerm ts us vs a x) -> SDTerm ts us vs b (App x x)

SPoly :: forall (ts :: [DSort]) (u :: DKind ts Kind) (uu :: NDKind ts Kind u) (us :: [DKind ts Kind]) (vs :: [DType ts us (Type :: DKind ts Kind)]) (a :: DType ts ((:) u us) (Type :: DKind ts Kind)) x x. (SingSing (NDKind ts Kind u) uu (WS x)) -> (SDTerm ts ((:) u us) (Map (ShiftSym ts us ((:) u us) u (Type :: DKind ts Kind) (InsZ :: Insert us ((:) u us) u)) vs) a x) -> SDTerm ts us vs (Pi uu a) (Poly x x)

SInst :: forall (ts :: [DSort]) (u :: DKind ts Kind) (uu :: NDKind ts Kind u) (us :: [DKind ts Kind]) (vs :: [DType ts us (Type :: DKind ts Kind)]) (b :: DType ts ((:) u us) (Type :: DKind ts Kind)) (a :: DType ts us u) x x x. (SingSing (NDKind ts Kind u) uu (WS x)) -> (SDTerm ts us vs (Pi uu b) x) -> (PolySing (NDType ts us u a) x) -> SDTerm ts us vs (Sub ts ((:) u us) us u (Type :: DKind ts Kind) (DelZ :: Delete ((:) u us) us u) a b) (Inst x x x)

SP :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (as :: [DType ts us (Type :: DKind ts Kind)]) (a :: DType ts us (Type :: DKind ts Kind)) (vs :: [DType ts us (Type :: DKind ts Kind)]) x x. (PolySing (Prim ts us as a) x) -> (PolySing (Prod (DTerm ts us vs) as) x) -> SDTerm ts us vs a (P x x)

SListLit :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DType ts us (Type :: DKind ts Kind)) (vs :: [DType ts us (Type :: DKind ts Kind)]) x x. (PolySing (NDType ts us (Type :: DKind ts Kind) a) x) -> (PolySing [DTerm ts us vs a] x) -> SDTerm ts us vs ((:$) (List :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) (ListLit x x)

SOptionalLit :: forall (ts :: [DSort]) (us :: [DKind ts Kind]) (a :: DType ts us (Type :: DKind ts Kind)) (vs :: [DType ts us (Type :: DKind ts Kind)]) x x. (PolySing (NDType ts us (Type :: DKind ts Kind) a) x) -> (PolySing (Maybe (DTerm ts us vs a)) x) -> SDTerm ts us vs ((:$) (Optional :: DType ts us ((:~>) (Type :: DKind ts Kind) (Type :: DKind ts Kind))) a) (OptionalLit x x)

data SAggType (k :: Type) (ls :: [Text]) (as :: [k]) (x :: AggType (k :: Type) (ls :: [Text]) (as :: [k])) where

Constructors

SATZ :: forall (k :: Type). SAggType k ('[] :: [Text]) ('[] :: [k]) ATZ

SATS :: forall (k :: Type) (l :: Text) (a :: k) (ls :: [Text]) (as :: [k]) x x x. (SingSing Text l (WS x)) -> (PolySing (WrappedSing k a) x) -> (SAggType k ls as x) -> SAggType k ((:) l ls) ((:) a as) (ATS x x x)

Singleton Functions

sShift :: SInsert us qs a ins -> SDType ts us b x -> SDType ts qs b (Shift ts us qs a b ins x)

sShift1 :: SDType ts us b x -> SDType ts (a ': us) b (Shift ts us (a ': us) a b InsZ x)

sSub :: SDelete us qs a del -> SDType ts qs a x -> SDType ts us b r -> SDType ts qs b (Sub ts us qs a b del x r)

sSub1 :: SDType ts us a x -> SDType ts (a ': us) b r -> SDType ts us b (Sub ts (a ': us) us a b DelZ x r)

skSub :: SDelete ts rs a del -> SDKind rs a x -> SDKind ts b r -> SDKind rs b (KSub ts rs a b del x r)

skSub1 :: SDKind ts a x -> SDKind (a ': ts) b r -> SDKind ts b (KSub (a ': ts) ts a b DelZ x r)

skNormalize :: SDKind ts a x -> SDKind ts a (KNormalize ts a x)

stNormalize :: SDType ts us a x -> SDType ts us a (TNormalize ts us a x)

Defunctionalization Symbols

data KShiftSym ts ps a b :: Insert ts ps a -> DKind ts b ~> DKind ps b

data ShiftSym ts us qs a b :: Insert us qs a -> DType ts us b ~> DType ts qs b

Util

type family Map (f :: a ~> b) (xs :: [a]) :: [b] where ...

Equations

Map f '[] = '[]

Map f (x ': xs) = (f @@ x) ': Map f xs

data MapSym (f :: a ~> b) :: [a] ~> [b]